Renamed class and folder structure on module.client and added feature to all gamemodes.
