+++ src/Module.Server/Modes/Battle/CrpgBattleGameMode.cs
@@ -18,6 +18,7 @@ using Crpg.Module.Common.ChatCommands;
 using TaleWorlds.MountAndBlade.Multiplayer.GauntletUI.Mission;
 using TaleWorlds.MountAndBlade.View.MissionViews.Order;
 using Crpg.Module.GUI;
+using Crpg.Module.GUI.AmmoQuiverChange;
 using Crpg.Module.GUI.Commander;
 using Crpg.Module.GUI.EndOfRound;
 using Crpg.Module.GUI.HudExtension;
@@ -89,6 +90,7 @@ internal class CrpgBattleGameMode : MissionBasedMultiplayerGameMode
             new MissionAgentContourControllerView(),
             MultiplayerViewCreator.CreateMissionKillNotificationUIHandler(),
             new CrpgHudExtensionHandler(),
+            new AmmoQuiverChangeMissionView(),
             MultiplayerViewCreator.CreateMultiplayerMissionDeathCardUIHandler(),
             ViewCreator.CreateOptionsUIHandler(),
             ViewCreator.CreateMissionMainAgentEquipDropView(mission),
@@ -144,6 +146,7 @@ internal class CrpgBattleGameMode : MissionBasedMultiplayerGameMode
                     new CrpgUserManagerClient(), // Needs to be loaded before the Client mission part.
                     new MultiplayerMissionAgentVisualSpawnComponent(), // expose method to spawn an agent
                     new CrpgCommanderBehaviorClient(),
+                    new AmmoQuiverChangeMissionBehaviorClient(),
 #endif
                     battleClient,
                     new MultiplayerTimerComponent(), // round timer
@@ -155,6 +158,7 @@ internal class CrpgBattleGameMode : MissionBasedMultiplayerGameMode
                     new MissionBoundaryCrossingHandler(), // kills agent out of mission boundaries
                     new MultiplayerPollComponent(), // poll logic to kick player, ban player, change game
                     new CrpgCommanderPollComponent(),
+                    new AmmoQuiverChangeComponent(),
                     new MissionOptionsComponent(),
                     new CrpgScoreboardComponent(_gameType switch
                             {
