+++ src/Module.Server/Modes/Conquest/CrpgConquestGameMode.cs
@@ -15,6 +15,7 @@ using Crpg.Module.Api;
 using Crpg.Module.Common.ChatCommands;
 #else
 using Crpg.Module.GUI;
+using Crpg.Module.GUI.AmmoQuiverChange;
 using Crpg.Module.GUI.Commander;
 using Crpg.Module.GUI.Conquest;
 using Crpg.Module.GUI.Spectator;
@@ -68,6 +69,7 @@ internal class CrpgConquestGameMode : MissionBasedMultiplayerGameMode
             new SpectatorHudUiHandler(),
             new WarmupHudUiHandler(),
             new ConquestHudUiHandler(),
+            new AmmoQuiverChangeMissionView(),
             MultiplayerViewCreator.CreateMultiplayerMissionDeathCardUIHandler(),
             ViewCreator.CreateOptionsUIHandler(),
             ViewCreator.CreateMissionMainAgentEquipDropView(mission),
@@ -113,6 +115,7 @@ internal class CrpgConquestGameMode : MissionBasedMultiplayerGameMode
                 new MultiplayerMissionAgentVisualSpawnComponent(),
                 new CrpgCommanderBehaviorClient(),
                 new CrpgRespawnTimerClient(),
+                new AmmoQuiverChangeMissionBehaviorClient(),
 #endif
                 warmupComponent,
                 new CrpgConquestClient(),
@@ -123,6 +126,7 @@ internal class CrpgConquestGameMode : MissionBasedMultiplayerGameMode
                 new MissionBoundaryCrossingHandler(),
                 new MultiplayerPollComponent(),
                 new CrpgCommanderPollComponent(),
+                new AmmoQuiverChangeComponent(),
                 new MultiplayerAdminComponent(),
                 notificationsComponent,
                 new MissionOptionsComponent(),
